Revert "refactor: 💡 Account 폴더구조 변경"
This reverts commit 648ee6c0a4edc101c21935808ca15ad5dc857f92.
