+++ backend/joalarm/src/main/java/com/cupid/joalarm/account/controller/EmojiController.java
@@ -1,5 +1,7 @@
-package com.cupid.joalarm.account.emoji;
+package com.cupid.joalarm.account.controller;
 
+import com.cupid.joalarm.account.dto.EmojiDto;
+import com.cupid.joalarm.account.service.EmojiS3Service;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
