+++ backend/joalarm/src/main/java/com/cupid/joalarm/account/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package com.cupid.joalarm.account.jwt;
+package com.cupid.joalarm.account.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
