+++ backend/joalarm/src/main/java/com/cupid/joalarm/account/service/AccountService.java
@@ -1,8 +1,8 @@
-package com.cupid.joalarm.account;
+package com.cupid.joalarm.account.service;
 
-import com.cupid.joalarm.account.AccountDto;
-import com.cupid.joalarm.account.Account;
-import com.cupid.joalarm.account.AccountRepository;
+import com.cupid.joalarm.account.dto.AccountDto;
+import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.account.repository.AccountRepository;
 import com.cupid.joalarm.school.School;
 import com.cupid.joalarm.school.SchoolRepository;
 import com.sun.jdi.request.DuplicateRequestException;
