+++ backend/joalarm/src/main/java/com/cupid/joalarm/account/service/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
-package com.cupid.joalarm.account;
+package com.cupid.joalarm.account.service;
 
-import com.cupid.joalarm.account.Account;
-import com.cupid.joalarm.account.AccountRepository;
+import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.account.repository.AccountRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
