+++ backend/joalarm/src/main/java/com/cupid/joalarm/account/service/EmojiS3Service.java
@@ -1,8 +1,9 @@
-package com.cupid.joalarm.account.emoji;
+package com.cupid.joalarm.account.service;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectListing;
 import com.amazonaws.services.s3.model.S3ObjectSummary;
+import com.cupid.joalarm.account.dto.EmojiDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
