+++ backend/joalarm/src/main/java/com/cupid/joalarm/accountChatroom/service/AccountChatroomService.java
@@ -1,7 +1,7 @@
 package com.cupid.joalarm.accountChatroom.service;
 
-import com.cupid.joalarm.account.Account;
-import com.cupid.joalarm.account.AccountRepository;
+import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.account.repository.AccountRepository;
 import com.cupid.joalarm.accountChatroom.dto.AccountChatroomDto;
 import com.cupid.joalarm.accountChatroom.entity.AccountChatroom;
 import com.cupid.joalarm.accountChatroom.entity.AccountChatroomEmbedded;
