+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/Feed.java
@@ -6,7 +6,7 @@ import com.cupid.joalarm.base.entity.BaseTimeEntity;
 import com.cupid.joalarm.feed.comment.Comment;
 import com.cupid.joalarm.feed.like.Like;
 import com.cupid.joalarm.feed.tag.Tag;
-import com.cupid.joalarm.account.Account;
+import com.cupid.joalarm.account.entity.Account;
 import lombok.*;
 
 import javax.persistence.*;
