+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/FeedService.java
@@ -17,8 +17,8 @@ import com.cupid.joalarm.feed.like.LikeRepository;
 import com.cupid.joalarm.feed.tag.Tag;
 import com.cupid.joalarm.feed.tag.TagDto;
 import com.cupid.joalarm.feed.tag.TagRepository;
-import com.cupid.joalarm.account.Account;
-import com.cupid.joalarm.account.AccountRepository;
+import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.account.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
