+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/childcomment/ChildComment.java
@@ -1,6 +1,6 @@
 package com.cupid.joalarm.feed.childcomment;
 
-import com.cupid.joalarm.account.Account;
+import com.cupid.joalarm.account.entity.Account;
 import com.cupid.joalarm.base.entity.BaseTimeEntity;
 import com.cupid.joalarm.feed.comment.Comment;
 import com.cupid.joalarm.feed.like.Like;
