+++ backend/joalarm/src/main/java/com/cupid/joalarm/feed/like/LikeRepository.java
@@ -1,7 +1,7 @@
 package com.cupid.joalarm.feed.like;
 
 import com.cupid.joalarm.feed.Feed;
-import com.cupid.joalarm.account.Account;
+import com.cupid.joalarm.account.entity.Account;
 import com.cupid.joalarm.feed.childcomment.ChildComment;
 import com.cupid.joalarm.feed.comment.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
