+++ backend/joalarm/src/main/java/com/cupid/joalarm/school/School.java
@@ -1,6 +1,6 @@
 package com.cupid.joalarm.school;
 
-import com.cupid.joalarm.account.Account;
+import com.cupid.joalarm.account.entity.Account;
 import com.cupid.joalarm.feed.Feed;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.*;
