+++ backend/joalarm/src/main/java/com/cupid/joalarm/school/SchoolService.java
@@ -1,5 +1,9 @@
 package com.cupid.joalarm.school;
 
+import com.cupid.joalarm.account.entity.Account;
+import com.cupid.joalarm.feed.Feed;
+import com.cupid.joalarm.feed.FeedDto;
+import com.cupid.joalarm.feed.tag.TagDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
